+++ src/Microsoft.Azure.WebJobs.Extension.Dapr/Bindings/DaprPublishAsyncCollector.cs
@@ -11,6 +11,7 @@ namespace Microsoft.Azure.WebJobs.Extension.Dapr
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.WebJobs;
+    using Microsoft.Azure.WebJobs.Extension.Dapr.Services;
 
     class DaprPublishAsyncCollector : IAsyncCollector<DaprPubSubEvent>
     {
