+++ src/Microsoft.Azure.WebJobs.Extension.Dapr/Services/DaprServiceClient.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License.
 // ------------------------------------------------------------
 
-namespace Microsoft.Azure.WebJobs.Extension.Dapr
+namespace Microsoft.Azure.WebJobs.Extension.Dapr.Services
 {
     using System;
     using System.Collections.Generic;
