+++ src/Microsoft.Azure.WebJobs.Extension.Dapr/Triggers/DaprBindingTriggerBindingProvider.cs
@@ -11,6 +11,8 @@ namespace Microsoft.Azure.WebJobs.Extension.Dapr
     using Microsoft.AspNetCore.Http;
     using Microsoft.AspNetCore.Routing;
     using Microsoft.Azure.WebJobs;
+    using Microsoft.Azure.WebJobs.Extension.Dapr.Services;
+    using Microsoft.Azure.WebJobs.Extension.Dapr.Utils;
     using Microsoft.Azure.WebJobs.Host.Executors;
     using Microsoft.Azure.WebJobs.Host.Triggers;
 
@@ -31,7 +33,7 @@ namespace Microsoft.Azure.WebJobs.Extension.Dapr
             var attribute = parameter.GetCustomAttribute<DaprBindingTriggerAttribute>(inherit: false);
             if (attribute == null)
             {
-                return Utils.NullTriggerBindingTask;
+                return BindingUtils.NullTriggerBindingTask;
             }
 
             string daprBindingName = TriggerHelper.ResolveTriggerName(parameter, this.nameResolver, attribute.BindingName);
