+++ src/Microsoft.Azure.WebJobs.Extension.Dapr/Utils/BindingUtils.cs
@@ -1,14 +1,14 @@
-﻿// ------------------------------------------------------------
+// ------------------------------------------------------------
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 // ------------------------------------------------------------
 
-namespace Microsoft.Azure.WebJobs.Extension.Dapr
+namespace Microsoft.Azure.WebJobs.Extension.Dapr.Utils
 {
     using System.Threading.Tasks;
     using Microsoft.Azure.WebJobs.Host.Triggers;
 
-    static class Utils
+    internal static class BindingUtils
     {
         public static readonly Task<ITriggerBinding?> NullTriggerBindingTask =
             Task.FromResult<ITriggerBinding?>(null);
